split environment check in two steps
