+++ packages/interop-client/src/api/client.ts
@@ -1,0 +1,6 @@
+import { Api } from "./gateway.models.js";
+import { config } from "../config/env.js";
+
+export const apiClient = new Api({
+  baseURL: config.gatewayUrl,
+});
