--- packages/interop-client/src/client-assertion.ts
@@ -1,7 +1,0 @@
-import { kmsClientBuilder } from "./security/kms.js";
-
-export async function generateClientAssertion() {
-  const kmsClient = kmsClientBuilder();
-  const assertion = await kmsClient.buildJWT();
-  return assertion;
-}
