--- packages/interop-client/src/client.ts
@@ -1,6 +1,0 @@
-import { Api } from "./api/gateway.models.js";
-import { config } from "./config/interop-client.config.js";
-
-export const apiClient = new Api({
-  baseURL: config.gatewayUrl,
-});
