+++ packages/interop-client/src/get-client-assertion.ts
@@ -1,0 +1,7 @@
+import { kmsClientBuilder } from "./security/kmsClientHandler.js";
+
+export async function getClientAssertion() {
+  const kmsClient = kmsClientBuilder();
+  const assertion = await kmsClient.buildJWT();
+  return assertion;
+}
