+++ packages/interop-client/src/index.ts
@@ -1,11 +1,11 @@
-import { generateClientAssertion } from "./client-assertion.js";
-import { config } from "./config/interop-client.config.js";
+import { getClientAssertion } from "./get-client-assertion.js";
+import { config } from "./config/env.js";
 import { obtainVoucher } from "./get-voucher.js";
 import { getAgreementByPurposeId } from "./getAgreementByPurposeId.js";
 
 export const getAgreementByPurpose = async (purposeId: string) => {
   try {
-    const clientAssertion = await generateClientAssertion();
+    const clientAssertion = await getClientAssertion();
 
     const voucher = await obtainVoucher(config, clientAssertion);
